Refactor all specContainer using chakra, mobile responsive - ui overhaul

+++ backend/api/api.py
@@ -16,13 +16,6 @@ import soundfile as sf
 parser = reqparse.RequestParser()
 
 class ApiHandler(Resource):
-   
-    # def get(self):
-    #     return {
-    #         'resultStatus': 'SUCCESS',
-    #         'message': "Backend connection successful!"
-    #     }
-
     
     def get(self, file_name):
         print(file_name)
@@ -31,7 +24,6 @@ class ApiHandler(Resource):
     # Using to accept audio file, create spectrogram, and return.
     def post(self):
         library_choice = request.form["library"]
-        print(library_choice)
 
         return_data = io.BytesIO()
 
@@ -44,25 +36,35 @@ class ApiHandler(Resource):
         return send_file(return_data, mimetype = 'image/jpg', as_attachment=True, download_name='SpectrogramVersion.jpg') # temp_file_prefix
 
 
-def handleLibrosaSpec(return_data, request):
-    audio_file = request.files["audio_file"]
-    n_fft_val = int(request.form["n_fft"])
-    win_val = int(request.form["win_val"])
+def retreiveSpectrogramData(request):
+    if request.form["default_file"] == "true":
+        file_name = "./backend/api/clackLaptop.wav"
+    else:
+        audio_file = request.files["audio_file"]
+        file_name = str(random.randint(0,100000)) + ".wav"
+        audio_file.save(file_name)
 
-    file_name = str(random.randint(0,100000)) + ".wav"
-    audio_file.save(file_name)
 
     path = os.fspath(file_name)
     y, sr = librosa.load(path)
-    os.remove(file_name)
+
+    if request.form["default_file"] != "true":
+        os.remove(file_name)
+
+    return y, sr
+
+
+def handleLibrosaSpec(return_data, request):
+    win_val = int(request.form["win_val"])
+    n_fft_val = int(request.form["n_fft"])
+
+    y, sr = retreiveSpectrogramData(request)
 
     mel_speccy = librosa.feature.melspectrogram(y=y, sr=sr, n_fft=n_fft_val, win_length=win_val)
     M_db = librosa.power_to_db(mel_speccy, ref=np.max)
 
     # p = plt.figure(num=None, figsize=(8, 6))
     p2 = plt.subplot(111)
-    # p3 = plt.axis('on')
-    # 
     if (str(request.form['axes'])) == "false":
         p4 = plt.subplots_adjust(left=0,right=1, bottom=0, top=1)
         p5 = librosa.display.specshow(M_db, sr=sr)
@@ -88,42 +90,13 @@ def handleLibrosaSpec(return_data, request):
     os.remove(file_path)
 
 
-
-
-# wav = wave.open(file_name, 'rb')
-# frames = wav.readFrames(-1)
-# sound_info = np.frombuffer(frames, 'int16')
-# frame_rate = wav.getframerate()
-# print(frame_rate)
-# wav.close()
-
 def handleMatplotlibSpec(return_data, request):
     # Read the wav file (mono)
     
     n_fft_val = int(request.form["n_fft"])
-    # win_val = int(request.form["win_val"])
-    print("tried this")
-    print(str(request.form['axes']))
 
-    if request.form["default_file"] == "true":
-        file_name = "./backend/api/clackLaptop.wav"
-    else:
-        audio_file = request.files["audio_file"]
-        file_name = str(random.randint(0,100000)) + ".wav"
-        audio_file.save(file_name)
+    y, sr = retreiveSpectrogramData(request)
 
-
-    path = os.fspath(file_name)
-    y, sr = librosa.load(path)
-    
-    # sf.write('tmp.wav', y, sr)
-    # wav_file = wave.open('tmp.wav','r')
-
-
-    #samplingFrequency, signalData = wav.read(path)
-    if request.form["default_file"] != "true":
-        os.remove(file_name)
-    
     # Plot the signal read from wav file
     # p = plt.subplot(211)
     # # p = plt.title('Spectrogram of a wav file with piano music')
